refactor: consolidate Input and SecretInput Storybook under Forms
Point forms/input at DesignSystem.Input, add secret_input story, and remove duplicate design_system copies.
